services/schwab-price-sync/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schwab-price-sync"
version = "0.1.0"
description = "One-shot Schwab price history sync worker"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.0",
  "schwab-py>=1.5.0",
  "SQLAlchemy>=2.0.0",
]

[project.scripts]
schwab-price-sync = "schwab_price_sync.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]